first battle promt
