+++ app/src/main/AndroidManifest.xml
@@ -15,9 +15,6 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <meta-data
-            android:name="kontakt.io.API_KEY"
-            android:value="@string/kontakt_io_api_key" />
 
         <activity android:name=".activity.MainActivity">
             <intent-filter>
