+++ app/src/main/java/com/baecon/rockpaperscissorsapp/activity/MainActivity.java
@@ -1,5 +1,7 @@
 package com.baecon.rockpaperscissorsapp.activity;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -9,9 +11,26 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.baecon.rockpaperscissorsapp.R;
+import com.baecon.rockpaperscissorsapp.rest.ApiClient;
+import com.baecon.rockpaperscissorsapp.rest.ApiInterface;
+import com.kontakt.sdk.android.ble.connection.OnServiceReadyListener;
+import com.kontakt.sdk.android.ble.manager.ProximityManager;
+import com.kontakt.sdk.android.ble.manager.ProximityManagerFactory;
+import com.kontakt.sdk.android.ble.manager.listeners.IBeaconListener;
+import com.kontakt.sdk.android.ble.manager.listeners.simple.SimpleIBeaconListener;
+import com.kontakt.sdk.android.common.KontaktSDK;
+import com.kontakt.sdk.android.common.profile.IBeaconDevice;
+import com.kontakt.sdk.android.common.profile.IBeaconRegion;
+
+import retrofit2.Call;
+
+import static java.lang.Boolean.TRUE;
 
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = MainActivity.class.getSimpleName();
+    private ProximityManager proximityManager;
+    private static final String APIKEY = "ok";
+
 
     SharedPreferences sharedPrefs;
 
@@ -40,19 +59,83 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-        // FLiegt raus
-        //TODO: Suche Beacon
-        // Wenn gefunden, dann validiere
-        // Wenn TRUE dann rufe Spiele Bildschirm auf
-        // Code aus GameActivity
-        ImageView background =  (ImageView) findViewById(R.id.background);
-        background.setOnClickListener(new View.OnClickListener() {
+        KontaktSDK.initialize(APIKEY);
+
+        proximityManager = ProximityManagerFactory.create(this);
+        proximityManager.setIBeaconListener(createIBeaconListener());
+
+
+    }
+
+    //TODO auslagern in eigene Klasse ?
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        startScanning();
+    }
+
+    @Override
+    protected void onStop() {
+        proximityManager.stopScanning();
+        super.onStop();
+    }
+
+    @Override
+    protected void onDestroy() {
+        proximityManager.disconnect();
+        proximityManager = null;
+        super.onDestroy();
+    }
+
+    private void startScanning() {
+        proximityManager.connect(new OnServiceReadyListener() {
             @Override
-            public void onClick(View v) {
-                Intent bluetoothIntent = new Intent(MainActivity.this, GameActivity.class);
-                startActivity(bluetoothIntent);
+            public void onServiceReady() {
+                proximityManager.startScanning();
             }
         });
+    }
+
+    private IBeaconListener createIBeaconListener() {
+        return new SimpleIBeaconListener() {
+            @Override
+            public void onIBeaconDiscovered(IBeaconDevice ibeacon, IBeaconRegion region) {
+                if (isValidBeacon(ibeacon.getUniqueId()) ) {
+                    //TODO starte spiel
+                    new AlertDialog.Builder(MainActivity.this)
+                            .setIcon(R.drawable.battleicon)
+                            .setTitle("Loot upassen")
+                            .setMessage("iBeacon " + ibeacon.getUniqueId() + " wurde gefunden.")
+                            .setPositiveButton(R.string.ok_label, new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which) {
+                                    Intent gameIntent = new Intent(MainActivity.this, GameActivity.class);
+                                    startActivity(gameIntent);
+                                }
+                            })
+                            .setNegativeButton(R.string.cancel_label, new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which) {
+
+                                }
+                            })
+                            .show();
+                }
+            }
+        };
+    }
+
+    private boolean isValidBeacon(String beaconID){
+        ApiInterface apiService = ApiClient.getClient().create(ApiInterface.class);
+
+        Call<String> call = apiService.isValidBeacon(beaconID);
+//        try {
+        return TRUE;
+//            return Boolean.valueOf(call.execute().body());
+//        } catch (IOException e) {
+//            return FALSE;
+//        }
     }
 
 }
